Refs/heads/all edit to angular backend

+++ Backend/Maen/src/main/java/com/triana/salesianos/dam/Maen/dto/electricityCompany/GetElectricityCompanyDTO.java
@@ -10,14 +10,14 @@ public record GetElectricityCompanyDTO(
         UUID id,
         String name,
         String urlImage,
-        int numOfContract
+        Long numOfContract
 ) {
     public static GetElectricityCompanyDTO of(ElectricityCompany ec){
         return GetElectricityCompanyDTO.builder()
                 .id(ec.getId())
                 .name(ec.getName())
                 .urlImage(ec.getLogotype())
-                .numOfContract(0)
+                .numOfContract(0L)
                 .build();
     }
 }
